+++ packages/block-editor/src/hooks/use-zoom-out.js
@@ -2,7 +2,7 @@
  * WordPress dependencies
  */
 import { useSelect, useDispatch } from '@wordpress/data';
-import { useEffect, useRef } from '@wordpress/element';
+import { useEffect } from '@wordpress/element';
 
 /**
  * Internal dependencies
@@ -11,37 +11,33 @@ import { store as blockEditorStore } from '../store';
 import { unlock } from '../lock-unlock';
 
 /**
- * A hook used to set the zoomed out view, invoking the hook sets the mode.
+ * A hook used to set the editor mode to zoomed out mode, invoking the hook sets the mode.
  *
- * @param {boolean} zoomOut If we should zoom out or not.
+ * @param {boolean} zoomOut If we should enter into zoomOut mode or not
  */
 export function useZoomOut( zoomOut = true ) {
-	const { setZoomLevel } = unlock( useDispatch( blockEditorStore ) );
+	const { setZoomLevel, resetZoomLevel } = unlock(
+		useDispatch( blockEditorStore )
+	);
 	const { isZoomOut } = unlock( useSelect( blockEditorStore ) );
 
-	const originalIsZoomOutRef = useRef( null );
-
 	useEffect( () => {
-		// Only set this on mount so we know what to return to when we unmount.
-		if ( ! originalIsZoomOutRef.current ) {
-			originalIsZoomOutRef.current = isZoomOut();
-		}
-
-		// The effect opens the zoom-out view if we want it open and the canvas is not currently zoomed-out.
-		if ( zoomOut && isZoomOut() === false ) {
-			setZoomLevel( 50 );
-		} else if (
-			! zoomOut &&
-			isZoomOut() &&
-			originalIsZoomOutRef.current !== isZoomOut()
-		) {
-			setZoomLevel( originalIsZoomOutRef.current ? 50 : 100 );
-		}
+		const isZoomOutOnMount = isZoomOut();
 
 		return () => {
-			if ( isZoomOut() && isZoomOut() !== originalIsZoomOutRef.current ) {
-				setZoomLevel( originalIsZoomOutRef.current ? 50 : 100 );
+			if ( isZoomOutOnMount ) {
+				setZoomLevel( 50 );
+			} else {
+				resetZoomLevel();
 			}
 		};
-	}, [ isZoomOut, setZoomLevel, zoomOut ] );
+	}, [] );
+
+	useEffect( () => {
+		if ( zoomOut ) {
+			setZoomLevel( 50 );
+		} else {
+			resetZoomLevel();
+		}
+	}, [ zoomOut, setZoomLevel, resetZoomLevel ] );
 }