+++ packages/edit-post/src/components/init-pattern-modal/index.js
@@ -27,8 +27,8 @@ export default function InitPatternModal() {
 			isNewPost: isCleanNewPost(),
 		};
 	}, [] );
-	const [ isModalOpen, setIsModalOpen ] = useState( () =>
-		isNewPost && postType === 'wp_block' ? true : false
+	const [ isModalOpen, setIsModalOpen ] = useState(
+		() => isNewPost && postType === 'wp_block'
 	);
 
 	if ( postType !== 'wp_block' || ! isNewPost ) {
@@ -82,8 +82,7 @@ export default function InitPatternModal() {
 							/>
 							<HStack justify="right">
 								<Button
-									// TODO: Switch to `true` (40px size) if possible
-									__next40pxDefaultSize={ false }
+									__next40pxDefaultSize
 									variant="primary"
 									type="submit"
 									disabled={ ! title }