+++ test/performance/specs/post-editor.spec.js
@@ -368,9 +368,11 @@ test.describe( 'Post Editor Performance', () => {
 			// Go to the test page.
 			await admin.editPost( draftId );
 			await perfUtils.disableAutosave();
-			const globalInserterToggle = page.getByRole( 'button', {
-				name: 'Toggle block inserter',
-			} );
+			const globalInserterToggle = page
+				.getByRole( 'region', { name: 'Editor top bar' } )
+				.getByRole( 'button', {
+					name: 'Block Inserter',
+				} );
 
 			const samples = 10;
 			const throwaway = 1;
@@ -424,9 +426,11 @@ test.describe( 'Post Editor Performance', () => {
 			// Go to the test page.
 			await admin.editPost( draftId );
 			await perfUtils.disableAutosave();
-			const globalInserterToggle = page.getByRole( 'button', {
-				name: 'Toggle block inserter',
-			} );
+			const globalInserterToggle = page
+				.getByRole( 'region', { name: 'Editor top bar' } )
+				.getByRole( 'button', {
+					name: 'Block Inserter',
+				} );
 			// Open Inserter.
 			await globalInserterToggle.click();
 
@@ -483,9 +487,11 @@ test.describe( 'Post Editor Performance', () => {
 			await admin.editPost( draftId );
 			await perfUtils.disableAutosave();
 
-			const globalInserterToggle = page.getByRole( 'button', {
-				name: 'Toggle block inserter',
-			} );
+			const globalInserterToggle = page
+				.getByRole( 'region', { name: 'Editor top bar' } )
+				.getByRole( 'button', {
+					name: 'Block Inserter',
+				} );
 			const paragraphBlockItem = page.locator(
 				'.block-editor-inserter__menu .editor-block-list-item-paragraph'
 			);
@@ -535,9 +541,11 @@ test.describe( 'Post Editor Performance', () => {
 		test( 'Run the test', async ( { page, admin, perfUtils } ) => {
 			await admin.createNewPost();
 			await perfUtils.disableAutosave();
-			const globalInserterToggle = page.getByRole( 'button', {
-				name: 'Toggle block inserter',
-			} );
+			const globalInserterToggle = page
+				.getByRole( 'region', { name: 'Editor top bar' } )
+				.getByRole( 'button', {
+					name: 'Block Inserter',
+				} );
 
 			const testPatterns = [
 				{